Adding changes requested.
